redirect auth() user on profile , follow , block , unfollow

+++ app/Http/Controllers/Usercontroller.php
@@ -20,6 +20,11 @@ class Usercontroller extends Controller
     }
 
     function gprof($username){
+        if(auth()->user()->username ==  $username){
+            return redirect()->route('users.profile');
+        }
+
+        
         $user=User::where('username' , $username)->get()->first();
         if(! $user->isBlockedBy(auth()->user())){
         return view("pages.gprof")->with('user' , $user);
@@ -43,6 +48,8 @@ class Usercontroller extends Controller
     }
     function show($id)
     {
+
+
     $user =User::where('id', $id)->get()->first();
     $posts = User::find($user['id'])->posts;
     return view("users.show",['user'=>$user,'posts'=>$posts]);
@@ -71,50 +78,6 @@ class Usercontroller extends Controller
 
 
 
-    function forgot_password_show(){
-        $page_title = "Reset password";
-        return view('dashboard.forgot_password',compact('page_title'));
-    }
-
-    function forgot_password_email_verify(){
-        $admin = Admin::where('email',request('email'))->first();
-        if(!empty($admin)){
-
-            $token = Password::broker('admin')->createToken($admin);
-            DB::table('password_resets')->insert([
-                'email'=>$admin->email,
-                'token'=>$token,
-                'created_at'=>Carbon::now()
-            ]);
-            Mail::to($admin->email)->send(new AdminResetPassword(['data'=>$admin,'token'=>$token]));
-            return redirect()->back()->with('success', 'Reset link is sent to your email address');
-
-         }else{
-            return redirect()->back()->withErrors(['email_is_wrong' => 'passwords.user']);
-        }
-        return back();
-    }
-    function forgot_password_token_verify($token){
-        $check_token =  DB::table('password_resets')->where('token',$token)->where('created_at','>',Carbon::now()->subHours(2))->first();
-         if(!empty($check_token)){
-            return view('dashboard.forgot_password_final',['data' => $token]);
-        }else{
-            return redirect(Route('reset_password_final'));
-        }
-    }
-    function forgot_password_final(UpdatePassword $request,$token){
-        $check_token =  DB::table('password_resets')->where('token',$token)->where('created_at','>',Carbon::now()->subHours(2))->first();
-        if(!empty($check_token)){
-            $admin = Admin::where('email',$check_token->email)->update([
-                'password' => bcrypt($request->input('password'))
-            ]);
-            DB::table('password_resets')->where('email',$check_token->email)->delete();
-            auth('admin')->attempt(['email'=>$check_token->email,'password'=>$request->input('password')]);
-            return redirect(Route('dashboard'));
-        }else{
-            return redirect(Route('forgot_password'));
-        }
-    }
 
 
     function search(Request $request){
@@ -149,6 +112,10 @@ class Usercontroller extends Controller
     }
 
     function unfollow(Request $request){
+
+        if(auth()->user()->id ==  $request->userid){
+            return redirect()->back();
+        }
 
         if($request->ajax()){
             $remove = User_follow::where('target_id',$request->userid)->where('user_id' , auth()->user()->id)->delete();
@@ -191,6 +158,10 @@ class Usercontroller extends Controller
 
     function follow(Request $request){
 
+        if(auth()->user()->id ==  $request->userid){
+            return redirect()->back();
+        }
+
         if($request->ajax()){
             $remove = User_follow::create([  'user_id' => auth()->user()->id  , 'target_id' => $request->userid]);
 
@@ -218,6 +189,9 @@ class Usercontroller extends Controller
 
     function block(Request $request){
 
+        if(auth()->user()->id ==  $request->userid){
+            return redirect()->back();
+        }
         if($request->ajax()){
             $remove = User_block::create([  'user_id' => auth()->user()->id  , 'target_id' => $request->userid]);
             $rem = User_follow::where('target_id',$request->userid)->where('user_id' , auth()->user()->id)->delete();
